+++ src/App.tsx
@@ -1,18 +1,18 @@
 import React from 'react';
 import Body from './Components/Body';
 import Nav from './Components/Nav';
-import { RecoilTest } from './Components/recoilTest';
+// import { RecoilTest } from './Components/recoilPOST';
 
 const App = () => {
-	return (
-		<>
-			<React.Suspense fallback={<div>loading</div>}>
-				<Nav />
-				<Body />
-				<RecoilTest />
-			</React.Suspense>
-		</>
-	);
+  return (
+    <>
+      <React.Suspense fallback={<div>loading</div>}>
+        <Nav />
+        <Body />
+        {/* <RecoilTest /> */}
+      </React.Suspense>
+    </>
+  );
 };
 
 export default App;
